requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.51.0
websockets==13.1
httpx==0.27.2
twilio==9.3.3
numpy>=1.26.4
audioop-lts>=0.2.1  # Python 3.13+ で audioop の代替（TBD-006対応）

# テスト用
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.2  # FastAPI TestClient (async)
